fuck yeah

+++ src/agent/task/mod.rs
@@ -11,7 +11,7 @@ pub trait Task: std::fmt::Debug {
     fn to_prompt(&self) -> Result<String>;
     fn get_functions(&self) -> Vec<Namespace>;
 
-    fn max_history_visibility(&self) -> usize {
+    fn max_history_visibility(&self) -> u16 {
         15
     }
